Task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	long long arr[50] = {1, 2};
+	long arr[50] = {1, 2};
 	int i;
 
 	for (i = 2; i <= 49 ; i++)
@@ -17,9 +17,9 @@ int main(void)
 	}
 	for (i = 0; i <= 48; i++)
 	{
-		printf("%lld, ", arr[i]);
+		printf("%ld, ", arr[i]);
 	}
-	printf("%lld", arr[49]);
+	printf("%ld", arr[49]);
 
 	return (0);
